assignment2-handout/pyproject.toml
[project]
name = "assignment2-handout"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.2.6",
    "open3d>=0.19.0",
    "opencv-python>=4.13.0.92",
    "plyfile>=1.1.3",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "scipy>=1.15.3",
]
